First public iteration
This project was made with heavy use of chatGPT to try and quickly send out a functional converter of the SliST system!
I take no credit for the code written.
I know HOW to code, but an unfamiliar with HTML and wished to get this done in the hour.
We have been using SliST prior to this script.

AI generation should not be monetized. It is a tool that should be reserved for expression by those who lack the ability to. We hope it doesn't deter you from using this system.